dags/requirements.txt
apache-airflow
pandas==2.1.4
SQLAlchemy
psycopg2-binary
pytest 
pytest-mock
clickhouse-connect